requirements.txt
requests>=2.31.0
flask>=2.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
folium>=0.14.0
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0